Investigated and found the source of my bugs
(1) My hashmap to check the properties in the properties.txt file doesnt work at all. This is because Java tries to compare an object and a string and doesnt do it too well. (2) my resource.toString() darling that I use all over doesnt translate into what I need in some/most cases. Going to have to write a function for this.